agents/requirements.txt
openai
deepeval
python-dotenv
pytest